+++ Public/Start-PhoenixRestoreJob.ps1
@@ -22,11 +22,15 @@ function Start-PhoenixRestoreJob {
         [Parameter()]
         [ValidateSet(
             'WinGet',
-            'Chocolatey'
+            'Chocolatey',
+            'Scoop',
+            'PowerShell Gallery'
         )]
         [string[]]$Provider = @(
             'WinGet',
-            'Chocolatey'
+            'Chocolatey',
+            'Scoop',
+            'PowerShell Gallery'
         ),
 
         [Parameter()]